refactor: split ansible logic with mixins
